src/requirements.txt
redis==5.2.1
asyncpraw==7.8.1
beautifulsoup4==4.13.3
fastapi==0.115.8
itunespy==1.6.1
pet_pet_gif==1.0.2
Pillow==11.1.0
playwright==1.50.0
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
uvicorn==0.34.0
numpy==2.2.2
pydantic==2.10.6
pygments==2.19.1 